requirements.txt
django
django-filter
djangorestframework
drf-yasg
markdown
pytest